fix: create asset location function
