+++ apps/storefront/src/index.d.ts
@@ -19,17 +19,6 @@ type ChannelPlatform =
   | 'wordpress'
   | 'custom';
 
-enum Environment {
-  Production = 'production',
-  Staging = 'staging',
-  Integration = 'integration',
-  Local = 'local',
-}
-
-type EnvSpecificConfig<ValueType> = {
-  [key in Environment]: ValueType;
-};
-
 declare interface Window {
   tipDispatch: import('@/shared/global/context/config.ts').DispatchProps;
   globalTipDispatch: any;
@@ -38,10 +27,11 @@ declare interface Window {
       channel_id: number;
       store_hash: string;
       platform: ChannelPlatform;
-      environment: Environment;
+      environment: import('@/types/global').Environment;
     };
   };
   b2b: {
+    __get_asset_location: (filename: string) => string;
     initializationEnvironment: import('./load-functions').InitializationEnvironment;
     callbacks: import('@/utils/b3CallbackManager').default;
     utils: {
