+++ apps/storefront/src/shared/service/request/base.ts
@@ -1,3 +1,5 @@
+import { Environment, EnvSpecificConfig } from '@/types';
+
 const ENVIRONMENT_B2B_API_URL: EnvSpecificConfig<string> = {
   local: import.meta.env.VITE_B2B_URL ?? 'http://localhost:9000',
   integration: 'https://api-b2b.integration.zone',
